Added a playing position tracking system
I've begun the process of making it possible for players to see what position they are in during the game. This is now working for the current player, but I still need to create a way of showing the other players positions as well. I've planned how I will do this but I have not yet finished implementing this.

Also patched a few bugs.