Added Eagle
I've implemented an eagle that patrols the sky above the level, seeking out exposed players and attacking them based on how long they've been exposed and how much in the lead they are (players in last place are never targetted). If the player gets out of sight, or the eagle collides with an obstacle on the attack, the attack will be called off and the eagle will return to patrolling.
I've tested this works in single player, I will test this properly in a lan multiplayer environment tomorrow.