Imported music, sounds, models, and cleaned folders pt 1
Moved Zak's models to correct directory, and exported fbx's, moving them into the project and deleting the misplaced obj files that should not have been in the project to begin with. Had to delete the obj models from the scene that Elliot had added earlier as they are no longer able to be used.
Cleaned out old files that we are no longer using in the game project, and organised the folders.
I also imported Elliot's music and sound files, though these have not yet been added to the game.
I replaced the level model with the new export from the work Max has been doing as well, and finally I made a few adjustments to the checkpoint light and added similar lights (minus the checkpoint scripts etc) to the trees to make it easier for players to know where to go in the level.