Fixed a few bugs and began to rework the movement code
I'm mostly done with the movement code, just need to keep playing with variable values to get the feel right. Animations are still a bit janky on the jump/glide so I'm going to keep working on it

I've fixed a few minor issues with the level and the lobby screen, and I've added Elliot's new shader to the in-game squirrel model. We'll probably have to remove the toon shader effect to stop the weird lighting bug we're getting.

I've got a few ideas for some adjustments to make tomorrow and Tuesday/Wednesday which I'll try to get to soon.