Improved UI, added a hint system, finished jump code
I've changed the font for some elements of the in game ui and rescaled it properly.
I've implemented a new feature - contextual text hints at the bottom of the screen that show up the first time the player encounters something new. I will add a setting to the options menu to never show these in the next few days.
I've finished reworking the player movement. Animations are still a little janky but I'm hoping to iron out the minor issues before the final hand in.
Don't forget to rebuild the lighting data before doing a build. The shadows go wrong otherwise.