Presentation Adjustments
Renamed the presentation so it was in line with our current naming convention.

Added in pictures where relevant and a gameplay gif for the animation slide (since a picture made no sense there).

Adjusted the colours on the graphs to make sure they were more clear.

Made small text changes here and there for clarity.

Adjusted some sizes of pictures so they looked better

Added the gif used into the group videos folder.