Added an in-game menu to adjust volume and quit
Quit option is only available to clients, if you are the server this does not work. I'll look into making the quit option available as the server if there only one player left later this afternoon. Just wanted to commit this fix now.
Also the end scoreboard "should" now be working correctly.